==> process steps/04_match_vms_fishtix.Rmd <==
---
title: "Match Fish Ticket to VMS Data"
output: html_document
---

## Purpose

This portion of the data processing matches the cleaned PacFIN fish tickets to the cleaned VMS data. We match the two datasets together using unique vessel identification numbers that are common to both datasets. We define a fishing "trip" as the collection of VMS observations directly preceding a landed fish ticket. Hence, all of the VMS pings that occur between two tickets are considered part of the latter ticket's trip. We enforce a 7-day "lookback window" (S. Jardine, *pers. comm.*) such that if there are more than 7 days [an option that can be changed] between successive tickets for the same vessel, we only include data for those 7 days.

Later, the data are filtered and cleaned for depth, unrealistic speeds between pings, and corrected for multiple-ticket days.

## Setup and Data Import

```{r, echo=FALSE}

# library(foreign)
library(rgdal)
library(rgeos)
library(tidyverse)
library(lubridate)
library(magrittr)
library(here)

# ggplot theme
plot_theme <-   theme_minimal()+
  theme(text=element_text(family="sans",size=12,color="black"),
        legend.text = element_text(size=14),
        axis.title=element_text(family="sans",size=14,color="black"),
        axis.text=element_text(family="sans",size=8,color="black"),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank())
theme_set(plot_theme)

rm(list=setdiff(ls(),c('process_year','alltime')))
```

Choose year to process

```{r}
# process_year <- 2012
```

Load fish ticket data (processed), including the last month of the previous year (for handling trips spanning January 1st of the current year)

```{r load tix}
fishtix <- read_rds(here::here('data','processed','fish tickets',paste0(process_year,'fishtix.rds')))

if(process_year>2009){
  fishtix_prev <- read_rds(here::here('data','processed','fish tickets',paste0(process_year-1,'fishtix.rds'))) %>% 
    filter(month(date)==max(month(date)))
  fishtix <- bind_rows(fishtix_prev,fishtix)
  rm(fishtix_prev)
}

```

Load VMS data, including the last month of the previous year (for handling trips spanning January 1st of the current year)

```{r load vms}
vms <- read_rds(here::here('data','processed','vms',paste0(process_year,'_vms_clean.rds')))

if(process_year>2009){
  vms_prev <- read_rds(here::here('data','processed','vms',paste0(process_year-1,'_vms_clean.rds'))) %>% 
    filter(month(UTCDATETIM)==max(month(UTCDATETIM)))
  vms <- bind_rows(vms_prev,vms)
  rm(vms_prev)
}
```

## Prep VMS Data for Matching

### Add West Coast Dates

PacFIN reports in west coast time, so create two columns with date/time in Pacific time zone

```{r}
vms %<>%
  # with time (hour, minute)
  mutate(westcoastdate = with_tz(UTCDATETIM, tzone = "America/Los_Angeles")) %>%
  # without time
  mutate(westcoastdate_notime = as_date(westcoastdate))
```

### Define Lookback Window and Begin/End Dates

Save the dates that serve as the beginning and end of that year's data

```{r begin end}
year_begin <- ymd(paste0(process_year,"-01-01"), tz= "America/Los_Angeles")
year_end <- ymd(paste0(process_year,"-12-31"), tz= "America/Los_Angeles")
```

Create lookback object. Here is where the lookback window duration can be changed. Default is 7 days.

```{r lookback}
lookback_window <- 7

if(process_year == 2009){
  lookback_begin <- year_begin
} else{
  lookback_begin <- year_begin - days(lookback_window)
}
```

## Prep Fish Ticket Data for Matching

Truncate fish ticket data to match VMS year, with the lookback window where appropriate. This should prevent fish tickets from the previous year being assigned the same VMS data points.

```{r define fishtix year}
fishtix %<>%
  filter(date >= lookback_begin & date <= year_end) %>% 
  
  #remove unneeded date columns
  select(Rec_ID:date,TARGET_lbs:species_code_all)
```

Add begin and end dates for each ticket to use in VMS matching. These define the windows of time in which we "search" for matching VMS pings. We also add a column for trip duration in days, for use in QA later.

```{r add begin end,warning=F}
ptm <- proc.time()
fishtix %<>%

  # group by vessel
  group_by(drvid) %>%
  
  #add month
  mutate(month=month(date)) %>% 
  
  # add a time for the end of day
  mutate(end = ymd_hms(paste(date, "23:59:59"),tz= "America/Los_Angeles")) %>%
  # arrange(date) %>%

  # find the ticket BEFORE each ticket and note its time. Instead of creating NAs for the first ticket of the year for each vessel, assign a default value of the first ticket minus the lookback window.
  mutate(prev_ticket_dttm=lag(end,order_by=end,default=first(end,order_by = end)-days(lookback_window))) %>%

  # add lookback window
  mutate(lookback_dttm=end-days(lookback_window)) %>%

  # choose "begin" date for each ticket as the LATEST in time between previous ticket date and lookback date
  mutate(begin=ifelse(lookback_dttm>prev_ticket_dttm,lookback_dttm,prev_ticket_dttm) %>% as_datetime(tz= "America/Los_Angeles")) %>%

  # add trip duration
  mutate(trip_dur=interval(begin,end)) %>% mutate(trip_dur=time_length(trip_dur,"days")) %>% 
  # select(-prev_ticket_dttm,-lookback_dttm) %>% 

  ungroup()
x<-proc.time()-ptm
cat('Defining lookback window for all data took',x[3]/60,'minutes to run.')
```

## Match VMS to Fish Tickets

Here is where we do the actual matching of VMS data to fish tickets. Matching is done by vessel ID number and time, which are recorded in both datasets. We first join all VMS data for each vessel, then filter such that only the records within the appropriate window of time for each ticket (defined in the previous step) are retained.

```{r match data, eval=TRUE}
ptm <- proc.time()

join_vms_fishtix <- function(mth){
  
  tmptix <- fishtix %>% 
    filter(month==mth)
  
  tmpvms <- vms %>% filter(month(westcoastdate)%in%c(mth-1,mth))
  
  tmptix %>%
    # Left-join the VMS data by Vessel ID (drvid/DOCNUM)
    left_join(tmpvms,by=c('drvid'='DOCNUM')) %>% 
    
    # filter the VMS data such that the data for each trip is correctly truncated to fall within the lookback window. 
    # If any tickets did NOT match VMS data, they will not have associated dates from VMS.
    # We retain these, and then produce a boolean indicator denoting that they did not match.
    
    filter((westcoastdate >= begin & westcoastdate <= end)| is.na(westcoastdate)) %>%
    mutate(has_vms=ifelse(is.na(westcoastdate),0,1)) %>% 
    ungroup()
}

fishtix_vms <- purrr::map_df(unique(fishtix$month),~join_vms_fishtix(.)) %>% 
  # finally, filter the edge case such that we remove trips ENDING before January 1st.
  # this will retain the VMS records that are before Jan 1 for any trips ending on or after Jan 1
    filter(year(date)==process_year)


## Join vms to fishtix
# fishtix_vms <- fishtix %>%
#   
#   # Left-join the VMS data by Vessel ID (drvid/DOCNUM)
#   left_join(vms,by=c('drvid'='DOCNUM')) %>% 
#   
#   # filter the VMS data such that the data for each trip is correctly truncated to fall within the lookback window. 
#   # If any tickets did NOT match VMS data, they will not have associated dates from VMS.
#   # We retain these, and then produce a boolean indicator denoting that they did not match.
#   
#   filter((westcoastdate >= begin & westcoastdate <= end)| is.na(westcoastdate)) %>%
#   mutate(has_vms=ifelse(is.na(westcoastdate),0,1)) %>% 
#   ungroup()
# 
x<-proc.time()-ptm
cat('Joining VMS data and filtering took',x[3]/60,'minutes to run for',process_year)

```

## Visualize Matching Results

We now have the full matched dataset, including retaining `r sum(!fishtix_vms$has_vms)` fish ticket records with no associated VMS data.

```{r view outputs}
# test<- fishtix_vms %>% 
#   select(Rec_ID,trip_dur,has_vms) %>%
#   mutate(trip_dur=as.integer(trip_dur)) %>% 
#   distinct() %>% 
#   ungroup() %>% 
#   mutate(has_vms=ifelse(has_vms,"VMS Trips","Non-VMS Trips"))
# How long are trips?
fishtix_vms %>% 
  dplyr::select(Rec_ID,trip_dur,has_vms) %>%
  mutate(trip_dur=as.integer(trip_dur)) %>% 
  distinct() %>% 
  ungroup() %>% 
  mutate(has_vms=ifelse(has_vms,"VMS Trips","Non-VMS Trips")) %>% 
  ggplot(aes(trip_dur,fill=has_vms))+
  geom_bar(position = 'dodge')+
  scale_x_discrete(limits=seq(0,7,by=1))+
  labs(x="Trip Duration (Days)",y="Number of Trips",title="Trip Duration",fill="")


# Count how many trips have multiple species (measured as multiple targets)
fishtix_vms %<>% mutate(multispecies=ifelse(nchar(species_code_all)>4,1,0))

fishtix_vms %>% 
  dplyr::select(Rec_ID,multispecies,has_vms) %>%
  distinct() %>% 
  mutate(has_vms=ifelse(has_vms,"VMS Trips","Non-VMS Trips")) %>% 
  ggplot(aes(factor(multispecies),fill=factor(multispecies)))+
  scale_fill_discrete(labels=c('No',"Yes"))+
  geom_bar()+
  facet_wrap(~has_vms)+
  labs(x="",y="Number of Fish Tickets",title="Number of Multispecies Trips",fill="Multispecies\nTrip")+
  theme(axis.text.x = element_blank())

# Distribution of number of trips per vessel, VMS vs. non-VMS records
fishtix_vms %>% 
  group_by(drvid,has_vms) %>% 
  summarise(ntix=length(unique(Rec_ID))) %>% 
  ungroup() %>% 
  mutate(has_vms=ifelse(has_vms,"VMS Vessels","Non-VMS Vessels")) %>% 
  ggplot(aes(ntix,..count..))+
  geom_histogram(bins=10)+
  facet_wrap(~has_vms)+
  labs(x=paste("Number of Trips in",process_year),y="Number of Vessels",title="Distribution of Number of Recorded Trips per Vessel")
```

## Save Results

Save a version that has all fish ticket data, including records with no VMS match, as well as a filtered dataset that excludes non-VMS records.

```{r}
fishtix_vms %<>% dplyr::select(-month)
fishtix_vms_all <- fishtix_vms
fishtix_vms_only <- fishtix_vms %>% filter(has_vms==1)
write_rds(fishtix_vms_all,here::here('data','processed','matched','matching',paste0(process_year,'matched_alltix.rds')))
write_rds(fishtix_vms_only,here::here('data','processed','matched','matching',paste0(process_year,'matched_vmstix_only.rds')))
```
